Crate convert_osm

Modules

elevation 🔒

extract 🔒

gtfs 🔒

parking 🔒

Structs

Options

Configures the creation of a RawMap from OSM and other input data. TODO Layering is now strange. Some of these are options are needed just for StreetNetwork, but many are the next level up and just for A/B Street’s convert_osm.

Enums

OnstreetParking

What roads will have on-street parking lanes? Data from https://wiki.openstreetmap.org/wiki/Key:parking:lane is always used if available.

PrivateOffstreetParking

If a building doesn’t have anything from public_offstreet_parking and isn’t tagged as a garage in OSM, how many private spots should it have?

PublicOffstreetParking

How many spots are available in public parking garages?

Functions

add_extra_buildings 🔒

bristol_hack 🔒

clip_map 🔒

convert

Create a RawMap from OSM and other input data.