Module convert_osm::osm_geom

Utilities for extracting concrete geometry from OSM objects.

Functions

get_multipolygon_members

glue_multipolygon
Take a bunch of partial PolyLines and attempt to glue them together into a single Ring. If the result isn’t complete and a separate boundary around the entire area is available, attempt to glue along that too. Note the result could be more than one disjoint polygon.

glue_to_boundary

multipoly_geometry