Struct convert_osm::Options

pub struct Options {
    pub map_config: MapConfig,
    pub onstreet_parking: OnstreetParking,
    pub public_offstreet_parking: PublicOffstreetParking,
    pub private_offstreet_parking: PrivateOffstreetParking,
    pub include_railroads: bool,
    pub extra_buildings: Option<String>,
    pub filter_crosswalks: bool,
    pub gtfs_url: Option<String>,
    pub elevation: bool,
}

Configures the creation of a RawMap from OSM and other input data. TODO Layering is now strange. Some of these are options are needed just for StreetNetwork, but many are the next level up and just for A/B Street’s convert_osm.

Fields
map_config: MapConfigonstreet_parking: OnstreetParkingpublic_offstreet_parking: PublicOffstreetParkingprivate_offstreet_parking: PrivateOffstreetParkinginclude_railroads: bool

OSM railway=rail will be included as light rail if so. Cosmetic only.

extra_buildings: Option<String>

If provided, read polygons from this GeoJSON file and add them to the RawMap as buildings.

filter_crosswalks: bool

Only include crosswalks that match a highway=crossing OSM node.

gtfs_url: Option<String>

Configure public transit using this URL to a static GTFS feed in .zip format.

elevation: bool
Implementations

impl Options

pub fn default_for_side(driving_side: DrivingSide) -> Options